6.5.y: Suse TW fixups  - Import Suse depmod/kmod patches - https://github.com/SUSE/kernel-source/tree/master/patches.rpmify  - Tweak mkspec to respect Suse /usr usage  - Run grub2-mkconfig post-install  - Get rid of the undesired test tweak in 0013-fedora-rpm.patch added with e0c74e5e1559d3b008ed4a790b9f4b7846a71844
We may have some additional work to do as my limited testing with Arch's defconfig + _kernel_on_diet="true" didn't boot. It might actually work with the full defconfig or using Suse's using _configfile="running-kernel". I'll run more tests as needed on my main machine as it was a PITA working on this with a dualcore broadwell laptop :frog: