Extra build pipelines for Fedora and Ubuntu (#864)
* ci: factorise kernel building code for Arch

* customization: non-arch: add config option to disable installing the kernel altogether

useful for ci

* customization: add toggle to disable use of 'script' command

needed for ci, otherwise yes "" | ./install.sh install just
freezes

* ci: add Ubuntu pipeline

* ci: add Fedora pipeline

* install.sh: git add changes in the work folder

Both RPM and DEB scripts rely on apparently, not only RPM

* ci: use github to retrieve sources

Probably faster since the runner are hosted by them ?

* ci: arch: update mirror list

Sometimes the ci fails because of "operation too slow"

* ci: follow CachyOS naming scheme for bore scheduler

'bore-eevdf' is just 'bore' now, which makes sense for those who know

* ci: make kernelconfig part of the release

Enables advanced users to check if the kernel they're downloading
has the correct config options. And we can use it to check if
everything is okay

* ci: make a release as long as one kernel builds

I think it's better than cancelling the whole thing.
Because sometimes we don't have time to fix
what's wrong with a distro for a while

* ci: build diet kernel

All good! back to diet kernel

* ci: arch: remove extra frag file

Let's check after a full build if it was really needed.

* config: copy over diet config from 6.6 to 6.7

* ci: #YOLO: build all combinations of distro and scheduler